requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
reportlab==4.0.4
qrcode==7.4.2
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.0.0
Flask-Session==0.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
